+++ src/features/contents/hooks/movie/useGetTopRatedMovies.ts
@@ -1,0 +1,9 @@
+import { useQuery } from '@tanstack/react-query';
+import { fetchTopRatedMovies } from 'entities/contents/model';
+
+export function useTopRatedMovies(page: number = 1) {
+  return useQuery({
+    queryKey: ['movies', 'top-rated', page],
+    queryFn: () => fetchTopRatedMovies(page),
+  });
+}
