+++ src/features/contents/hooks/tvs/useGetPopularTVs.ts
@@ -1,0 +1,9 @@
+import { useQuery } from '@tanstack/react-query';
+import { fetchPopularTVs } from 'entities/contents/model';
+
+export function usePopularTVs(page: number = 1) {
+  return useQuery({
+    queryKey: ['tvs', 'popular', page],
+    queryFn: () => fetchPopularTVs(page),
+  });
+}
