+++ src/features/contents/hooks/tvs/useGetTopRatedTVs.ts
@@ -1,0 +1,9 @@
+import { useQuery } from '@tanstack/react-query';
+import { fetchTopRatedTVs } from 'entities/contents/model';
+
+export function useTopRatedTVs(page: number = 1) {
+  return useQuery({
+    queryKey: ['tvs', 'top-rated', page],
+    queryFn: () => fetchTopRatedTVs(page),
+  });
+}
